+++ tibiakt-core/src/main/kotlin/com/galarzaa/tibiakt/core/parsers/HighscoresParser.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2022 Allan Galarza
+ * Copyright © 2023 Allan Galarza
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -64,7 +64,7 @@ public object HighscoresParser : Parser<Highscores?> {
                 ?: throw ParsingException("Could not find last update label")
             numericMatch.find(lastUpdateText)?.also {
                 val minutes = it.groups[0]!!.value.toInt()
-                lastUpdate = Clock.System.now().minus(minutes.minutes)
+                lastUpdated = Clock.System.now().minus(minutes.minutes)
             }
             val paginationData = boxContent.selectFirst("small")?.parsePagination()
                 ?: throw ParsingException("could not find pagination block")