+++ tibiakt-core/src/test/kotlin/com/galarzaa/tibiakt/core/parsers/LeaderboardParserTests.kt
@@ -34,7 +34,7 @@ class LeaderboardParserTests : FunSpec({
 
         leaderboard.shouldBeInstanceOf<Leaderboard>()
         with(leaderboard) {
-            rotation.current shouldBe true
+            rotation.isCurrent shouldBe true
             lastUpdated shouldNotBe null
             entries.shouldNotBeEmpty()
         }